+++ engine/src/main/java/dev/gamekit/components/Collider.java
@@ -19,10 +19,10 @@ public abstract class Collider extends Component {
   public static boolean DEBUG_DRAW = false;
   private static final Stroke SENSOR_DEBUG_STROKE =
     new BasicStroke(1, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL, 0, new float[]{ 10, 2 }, 0);
-  protected final ColliderFixture fixture;
+  protected final RefFixture fixture;
   protected Physics.CollisionListener collisionListener;
 
-  protected Collider(ColliderFixture fixture) {
+  protected Collider(RefFixture fixture) {
     this.fixture = fixture;
     fixture.setCollider(this);
   }
@@ -141,14 +141,14 @@ public abstract class Collider extends Component {
     }
   }
 
-  /** {@link ColliderFixture} extends {@link BodyFixture} adding a reference to its parent {@link Collider} */
-  public static class ColliderFixture extends BodyFixture {
+  /** {@link RefFixture} extends {@link BodyFixture} adding a reference to its parent {@link Collider} */
+  public static class RefFixture extends BodyFixture {
     public final String id = UUID.randomUUID().toString();
 
     private Collider collider;
 
-    /** Creates a {@link ColliderFixture} with the associated {@link Convex shape} */
-    public ColliderFixture(Convex shape) {
+    /** Creates a {@link RefFixture} with the associated {@link Convex shape} */
+    public RefFixture(Convex shape) {
       super(shape);
     }